옵션 저장 API 완료

+++ custom-cake-backend/src/main/kotlin/com/cake/customcakebackend/application/port/in/CakeOptionManagementUseCase.kt
@@ -9,7 +9,7 @@ interface CakeOptionManagementUseCase {
     fun loadCakeOptionInfo(cakeOptionType: Int, cakeOptionId: Long): CakeOption
     fun loadAllCakeOptionList(storeId: Long): Map<Int, List<CakeOption>>
     fun loadCakeOptionList(storeId: Long, cakeOptionType: Int): List<CakeOption>
-    fun saveCakeOption(storeId: Long, cakeOptionType: Int, cakeOptionRequest: CakeOptionRequest): Pair<Long, Long>  // type, id
+    fun saveCakeOption(storeId: Long, cakeOptionType: Int, cakeOptionRequest: CakeOptionRequest): Pair<Int, Long>  // type, id
     fun modifyCakeOption(): Pair<Long, Long>  // type, id
     fun deleteCakeOption()
 }
